Renamed cli to armasec_cli to fix ambiguity and added package to manifest
