+++ tests/armasec_cli/check.py
