+++ tests/armasec_cli/test_config.py
@@ -5,9 +5,9 @@ from pathlib import Path
 import pytest
 from typer import Context
 
-from cli.exceptions import Abort
-from cli.schemas import CliContext
-from cli.config import init_settings, attach_settings, dump_settings
+from armasec_cli.exceptions import Abort
+from armasec_cli.schemas import CliContext
+from armasec_cli.config import init_settings, attach_settings, dump_settings
 
 
 def test_init_settings__success():
@@ -28,7 +28,7 @@ def test_init_settings__raises_Abort_on_invalid_config():
 
 def test_dump_settings__success(tmp_path, mocker):
     dummy_settings_path = tmp_path / "settings.json"
-    with mocker.patch("cli.config.settings_path", new=dummy_settings_path):
+    with mocker.patch("armasec_cli.config.settings_path", new=dummy_settings_path):
         dump_settings(
             init_settings(
                 oidc_domain="test.domain",
@@ -45,7 +45,7 @@ def test_dump_settings__success(tmp_path, mocker):
 
 def test_attach_settings__success(tmp_path, mocker, mock_context):
     dummy_settings_path = tmp_path / "settings.json"
-    with mocker.patch("cli.config.settings_path", new=dummy_settings_path):
+    with mocker.patch("armasec_cli.config.settings_path", new=dummy_settings_path):
         dump_settings(
             init_settings(
                 oidc_domain="test.domain",
@@ -64,7 +64,7 @@ def test_attach_settings__success(tmp_path, mocker, mock_context):
 
 
 def test_attach_settings__raises_Abort_if_settings_file_is_not_found(mocker):
-    with mocker.patch("cli.config.settings_path", new=Path("fake-path")):
+    with mocker.patch("armasec_cli.config.settings_path", new=Path("fake-path")):
 
         @attach_settings
         def _helper(*_):
